+++ src/app/common/table-available-actions.ts
@@ -1,10 +1,39 @@
 import { FormBuilder, Validators, FormArray, FormGroup, FormControl} from '@angular/forms';
 import { ValidationService } from './custom-validation/validation.service'
+import { IMultiSelectSettings } from './multiselect-dropdown'
 
 export class AvailableTableActions {
 
   //AvailableOptions result depeding on component type
   public availableOptions : Array<any>;
+  private single_select: IMultiSelectSettings = {
+    pullRight: false,
+    enableSearch: true,
+    checkedStyle: 'glyphicon',
+    buttonClasses: 'btn btn-default',
+    selectionLimit: 0,
+    closeOnSelect: false,
+    showCheckAll: true,
+    showUncheckAll: true,
+    dynamicTitleMaxItems: 3,
+    maxHeight: '400px',
+    singleSelect: true,
+    allowCustomItem: false
+  };
+  private multi_select: IMultiSelectSettings = {
+    pullRight: false,
+    enableSearch: true,
+    checkedStyle: 'glyphicon',
+    buttonClasses: 'btn btn-default',
+    selectionLimit: 0,
+    closeOnSelect: false,
+    showCheckAll: true,
+    showUncheckAll: true,
+    dynamicTitleMaxItems: 3,
+    maxHeight: '400px',
+    singleSelect: false,
+    allowCustomItem: false
+  };
 
   // type can be : device,...
   // data is the passed extraData when declaring AvailableTableActions on each component
@@ -13,7 +42,7 @@ export class AvailableTableActions {
       case 'kapacitor-component':
         return this.getKapacitorAvailableActions();
         case 'alert-component':
-        return this.getAlertAvailableActions();
+        return this.getAlertAvailableActions(data);
         case 'alerteventhist-component':
         return this.getAlertEventHistAvailableActions();
         case 'alertevent-component':
@@ -51,7 +80,11 @@ export class AvailableTableActions {
 
   getAlertAvailableActions (data ? : any) : any {
     let tableAvailableActions = [
-    //Remove Action
+      //Deploy Action
+      {'title': 'Deploy', 'content' :
+        {'type' : 'button','action' : 'DeployAllSelected'}
+      },
+      //Remove Action
       {'title': 'Remove', 'content' :
         {'type' : 'button','action' : 'RemoveAllSelected'}
       },
@@ -60,8 +93,61 @@ export class AvailableTableActions {
         {'type' : 'selector', 'action' : 'ChangeProperty', 'options' : [
           {'title' : 'Active', 'type':'boolean', 'options' : [
             'true','false']
+          },
+          {'title': 'InfluxFilter','type':'input', 'options':
+            new FormGroup({
+              formControl : new FormControl('')
+            })
+          },
+          {'title': 'AlertNotify','type':'input', 'options':
+            new FormGroup({
+              formControl : new FormControl('', ValidationService.uintegerValidator)
+            })
+          },
+          {'title': 'GrafanaServer','type':'input', 'options':
+            new FormGroup({
+              formControl : new FormControl('')
+            })
+          },
+          {'title': 'GrafanaDashLabel','type':'input', 'options':
+            new FormGroup({
+              formControl : new FormControl('')
+            })
+          },
+          {'title': 'GrafanaDashPanelID','type':'input', 'options':
+            new FormGroup({
+              formControl : new FormControl('')
+            })
+          },
+          {'title': 'ExtraLabel','type':'input', 'options':
+            new FormGroup({
+              formControl : new FormControl('')
+            })
+          },
+          {'title': 'ExtraTag','type':'input', 'options':
+            new FormGroup({
+              formControl : new FormControl('')
+            })
+          },
+          {'title' : 'Endpoint', 'type':'multiselector', 'options' :
+            data[0], 'settings' : this.multi_select
+          },
+          {'title' : 'KapacitorID', 'type':'multiselector', 'options' :
+            data[1], 'settings' : this.single_select
+          },
+          {'title' : 'OperationID', 'type':'multiselector', 'options' :
+            data[2], 'settings' : this.single_select
           }
         ]},
+      },
+      //AppendProperty
+      {'title': 'Append property', 'content' :
+        {'type' : 'selector', 'action' : 'AppendProperty', 'options' : [
+          {'title' : 'Endpoint', 'type':'multiselector', 'options' :
+            data[0], 'settings' : this.multi_select
+          }
+          ]
+        }
       }
     ];
     return tableAvailableActions;
@@ -98,7 +184,10 @@ export class AvailableTableActions {
         {'type' : 'selector', 'action' : 'ChangeProperty', 'options' : [
           {'title' : 'Active', 'type':'boolean', 'options' : [
             'true','false']
-          }
+          },
+          {'title' : 'ExceptionID', 'type':'selector', 'options' : [
+            '-1','0','1','2']
+          },
         ]},
       }
     ];
@@ -126,6 +215,37 @@ export class AvailableTableActions {
     //Remove Action
       {'title': 'Remove', 'content' :
         {'type' : 'button','action' : 'RemoveAllSelected'}
+      },
+      // Change Property Action
+      {'title': 'Change property', 'content' :
+        {'type' : 'selector', 'action' : 'ChangeProperty', 'options' : [
+          {'title': 'AlertGroups','type':'input', 'options':
+            new FormGroup({
+              formControl : new FormControl('')
+            })
+          },
+          {'title': 'FieldResolutions','type':'input', 'options':
+            new FormGroup({
+              formControl : new FormControl('')
+            })
+          },
+        ]},
+      },
+      //AppendProperty
+      {'title': 'Append property', 'content' :
+        {'type' : 'selector', 'action' : 'AppendProperty', 'options' : [
+          {'title': 'AlertGroups','type':'input', 'options':
+            new FormGroup({
+              formControl : new FormControl('')
+            })
+          },
+          {'title': 'FieldResolutions','type':'input', 'options':
+            new FormGroup({
+              formControl : new FormControl('', ValidationService.durationValidator)
+            })
+          },
+          ]
+        }
       }
     ];
     return tableAvailableActions;
@@ -144,13 +264,37 @@ export class AvailableTableActions {
     //Remove Action
       {'title': 'Remove', 'content' :
         {'type' : 'button','action' : 'RemoveAllSelected'}
+      },
+      // Change Property Action
+      {'title': 'Change property', 'content' :
+        {'type' : 'selector', 'action' : 'ChangeProperty', 'options' : [
+          {'title': 'MinHour','type':'input', 'options':
+            new FormGroup({
+              formControl : new FormControl('', Validators.compose([Validators.required, ValidationService.hourValidator]))
+            })
+          },
+          {'title': 'MaxHour','type':'input', 'options':
+            new FormGroup({
+              formControl : new FormControl('', Validators.compose([Validators.required, ValidationService.hourValidator]))
+            })
+          },
+          {'title': 'WeekDays','type':'input', 'options':
+            new FormGroup({
+              formControl : new FormControl('', Validators.compose([Validators.required, ValidationService.weekdaysValidator]))
+            })
+          }
+        ]},
       }
     ];
     return tableAvailableActions;
   }
   getTemplateAvailableActions (data ? : any) : any {
     let tableAvailableActions = [
-    //Remove Action
+      //Deploy Action
+      {'title': 'Deploy', 'content' :
+        {'type' : 'button','action' : 'DeployAllSelected'}
+      },
+      //Remove Action
       {'title': 'Remove', 'content' :
         {'type' : 'button','action' : 'RemoveAllSelected'}
       }