+++ src/app/kapacitor/kapacitor.data.ts
@@ -3,7 +3,8 @@ export const KapacitorComponentConfig: any =
     'name' : 'Kapacitor Backends',
     'table-columns' : [
       { 'title': 'ID', 'name': 'ID' },
-      { 'title': 'URL', 'name': 'URL' }
+      { 'title': 'URL', 'name': 'URL' },
+      { 'title': 'Imported', 'name': 'Imported','transform':'datetime' },
     ],
     'slug' : 'kapacitorcfg'
   };